Move url query related hooks to own folder
