+++ ui/src/hooks/index.ts
@@ -8,14 +8,13 @@ export * from './search/useSearchQueryParser';
 export * from './search/useSearchResults';
 export * from './stops';
 export * from './ui';
+export * from './urlQuery';
 export * from './useAsyncQuery';
 export * from './useCheckValidityAndPriorityConflicts';
 export * from './useChooseLineDropdown';
 export * from './useChooseRouteDropdown';
 export * from './useExtractRouteFromFeature';
 export * from './useFilterStops';
-export * from './useMapOpenQueryParam';
-export * from './useObservationDateQueryParam';
+export * from './usePagination';
 export * from './useShowRoutesOnModal';
-export * from './useUrlQuery';
 export * from './via';
